refactor(index): extract shared logic into indexapi_common and remove dead code from v1
